requirements.txt
dataclasses==0.6
Django==2.2
django-cors-headers==2.5.2
django-stubs==0.11.1
djangorestframework==3.9.2
djangorestframework-stubs==0.4.2
mypy==0.700
mypy-extensions==0.4.1
python-decouple==3.1
pytz==2019.1
sqlparse==0.3.0
typed-ast==1.3.1
typing-extensions==3.7.2
